requirements.txt
streamlit
spacy
nltk
scikit-learn
pandas
numpy
PyMuPDF
python-docx